Major bug fix and other things
Bug fix

Upon initial addition to the uptime feature I did not realize that the whole app was re-rendering every second. This was happening because the useEffect for the uptime was placed in this file, so the rest of the file tree was being updated when it did not need to.

Other things

Created multiple comps to render dynamcally styled terminals